fixing a bug in the getRandomSymOp function in SpaceGroupOps

+++ Source/OrientationLib/SpaceGroupOps/CubicLowOps.cpp
@@ -420,7 +420,7 @@ FOrientArrayType CubicLowOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;;
 
   init[0] = Detail::CubicLowDim1InitValue;
@@ -429,9 +429,9 @@ FOrientArrayType CubicLowOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = Detail::CubicLowDim1StepValue;
   step[1] = Detail::CubicLowDim2StepValue;
   step[2] = Detail::CubicLowDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -471,7 +471,7 @@ FOrientArrayType CubicLowOps::determineRodriguesVector(uint64_t seed, int choose
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::CubicLowDim1InitValue;
@@ -480,9 +480,9 @@ FOrientArrayType CubicLowOps::determineRodriguesVector(uint64_t seed, int choose
   step[0] = Detail::CubicLowDim1StepValue;
   step[1] = Detail::CubicLowDim2StepValue;
   step[2] = Detail::CubicLowDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
