+++ Source/OrientationLib/SpaceGroupOps/HexagonalLowOps.cpp
@@ -357,7 +357,7 @@ FOrientArrayType HexagonalLowOps::determineEulerAngles(uint64_t seed, int choose
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = HexDim1InitValue;
@@ -366,9 +366,9 @@ FOrientArrayType HexagonalLowOps::determineEulerAngles(uint64_t seed, int choose
   step[0] = HexDim1StepValue;
   step[1] = HexDim2StepValue;
   step[2] = HexDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -408,7 +408,7 @@ FOrientArrayType HexagonalLowOps::determineRodriguesVector(uint64_t seed, int ch
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = HexDim1InitValue;
@@ -417,9 +417,9 @@ FOrientArrayType HexagonalLowOps::determineRodriguesVector(uint64_t seed, int ch
   step[0] = HexDim1StepValue;
   step[1] = HexDim2StepValue;
   step[2] = HexDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
