+++ Source/OrientationLib/SpaceGroupOps/HexagonalOps.cpp
@@ -397,7 +397,7 @@ FOrientArrayType HexagonalOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = HexDim1InitValue;
@@ -406,9 +406,9 @@ FOrientArrayType HexagonalOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = HexDim1StepValue;
   step[1] = HexDim2StepValue;
   step[2] = HexDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -448,7 +448,7 @@ FOrientArrayType HexagonalOps::determineRodriguesVector(uint64_t seed, int choos
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = HexDim1InitValue;
@@ -457,9 +457,9 @@ FOrientArrayType HexagonalOps::determineRodriguesVector(uint64_t seed, int choos
   step[0] = HexDim1StepValue;
   step[1] = HexDim2StepValue;
   step[2] = HexDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
