+++ Source/OrientationLib/SpaceGroupOps/MonoclinicOps.cpp
@@ -284,7 +284,7 @@ FOrientArrayType MonoclinicOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::MonoclinicDim1InitValue;
@@ -293,9 +293,9 @@ FOrientArrayType MonoclinicOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = Detail::MonoclinicDim1StepValue;
   step[1] = Detail::MonoclinicDim2StepValue;
   step[2] = Detail::MonoclinicDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 36);
-  phi[2] = static_cast<float>(choose / (72 * 36));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 36);
+  phi[2] = static_cast<int32_t>(choose / (72 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -336,7 +336,7 @@ FOrientArrayType MonoclinicOps::determineRodriguesVector(uint64_t seed, int choo
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::MonoclinicDim1InitValue;
@@ -345,9 +345,9 @@ FOrientArrayType MonoclinicOps::determineRodriguesVector(uint64_t seed, int choo
   step[0] = Detail::MonoclinicDim1StepValue;
   step[1] = Detail::MonoclinicDim2StepValue;
   step[2] = Detail::MonoclinicDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 36);
-  phi[2] = static_cast<float>(choose / (72 * 36));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 36);
+  phi[2] = static_cast<int32_t>(choose / (72 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
