+++ Source/OrientationLib/SpaceGroupOps/OrthoRhombicOps.cpp
@@ -307,7 +307,7 @@ FOrientArrayType OrthoRhombicOps::determineEulerAngles(uint64_t seed, int choose
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = OrthoDim1InitValue;
@@ -316,9 +316,9 @@ FOrientArrayType OrthoRhombicOps::determineEulerAngles(uint64_t seed, int choose
   step[0] = OrthoDim1StepValue;
   step[1] = OrthoDim2StepValue;
   step[2] = OrthoDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -358,7 +358,7 @@ FOrientArrayType OrthoRhombicOps::determineRodriguesVector(uint64_t seed, int ch
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = OrthoDim1InitValue;
@@ -367,9 +367,9 @@ FOrientArrayType OrthoRhombicOps::determineRodriguesVector(uint64_t seed, int ch
   step[0] = OrthoDim1StepValue;
   step[1] = OrthoDim2StepValue;
   step[2] = OrthoDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
