+++ Source/OrientationLib/SpaceGroupOps/SpaceGroupOps.h
@@ -218,7 +218,7 @@ class OrientationLib_EXPORT SpaceGroupOps
     void _calcQuatNearestOrigin(const QuatF quatsym[24], int numsym, QuatF& qr);
 
     int _calcMisoBin(float dim[3], float bins[3], float step[3], const FOrientArrayType& homochoric);
-    void _calcDetermineHomochoricValues(uint64_t seed, float init[3], float step[3], float phi[3], int choose, float& r1, float& r2, float& r3);
+    void _calcDetermineHomochoricValues(uint64_t seed, float init[3], float step[3], int32_t phi[3], int choose, float& r1, float& r2, float& r3);
     int _calcODFBin(float dim[3], float bins[3], float step[3], FOrientArrayType homochoric);
 
   private:
