+++ Source/OrientationLib/SpaceGroupOps/TetragonalLowOps.cpp
@@ -300,7 +300,7 @@ FOrientArrayType TetragonalLowOps::determineEulerAngles(uint64_t seed, int choos
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::TetraDim1InitValue;
@@ -309,9 +309,9 @@ FOrientArrayType TetragonalLowOps::determineEulerAngles(uint64_t seed, int choos
   step[0] = Detail::TetraDim1StepValue;
   step[1] = Detail::TetraDim2StepValue;
   step[2] = Detail::TetraDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
@@ -352,7 +352,7 @@ FOrientArrayType TetragonalLowOps::determineRodriguesVector(uint64_t seed, int c
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::TetraDim1InitValue;
@@ -361,9 +361,9 @@ FOrientArrayType TetragonalLowOps::determineRodriguesVector(uint64_t seed, int c
   step[0] = Detail::TetraDim1StepValue;
   step[1] = Detail::TetraDim2StepValue;
   step[2] = Detail::TetraDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
