+++ Source/OrientationLib/SpaceGroupOps/TetragonalOps.cpp
@@ -328,7 +328,7 @@ FOrientArrayType TetragonalOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TetraDim1InitValue;
@@ -337,9 +337,9 @@ FOrientArrayType TetragonalOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = TetraDim1StepValue;
   step[1] = TetraDim2StepValue;
   step[2] = TetraDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -380,7 +380,7 @@ FOrientArrayType TetragonalOps::determineRodriguesVector(uint64_t seed, int choo
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TetraDim1InitValue;
@@ -389,9 +389,9 @@ FOrientArrayType TetragonalOps::determineRodriguesVector(uint64_t seed, int choo
   step[0] = TetraDim1StepValue;
   step[1] = TetraDim2StepValue;
   step[2] = TetraDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
