+++ Source/OrientationLib/SpaceGroupOps/TriclinicOps.cpp
@@ -281,7 +281,7 @@ FOrientArrayType TriclinicOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::TriclinicDim1InitValue;
@@ -290,9 +290,9 @@ FOrientArrayType TriclinicOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = Detail::TriclinicDim1StepValue;
   step[1] = Detail::TriclinicDim2StepValue;
   step[2] = Detail::TriclinicDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -332,7 +332,7 @@ FOrientArrayType TriclinicOps::determineRodriguesVector(uint64_t seed, int choos
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = Detail::TriclinicDim1InitValue;
@@ -341,9 +341,9 @@ FOrientArrayType TriclinicOps::determineRodriguesVector(uint64_t seed, int choos
   step[0] = Detail::TriclinicDim1StepValue;
   step[1] = Detail::TriclinicDim2StepValue;
   step[2] = Detail::TriclinicDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
