+++ Source/OrientationLib/SpaceGroupOps/TrigonalLowOps.cpp
@@ -320,7 +320,7 @@ FOrientArrayType TrigonalLowOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TrigDim1InitValue;
@@ -329,9 +329,9 @@ FOrientArrayType TrigonalLowOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = TrigDim1StepValue;
   step[1] = TrigDim2StepValue;
   step[2] = TrigDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -371,7 +371,7 @@ FOrientArrayType TrigonalLowOps::determineRodriguesVector(uint64_t seed, int cho
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TrigDim1InitValue;
@@ -380,9 +380,9 @@ FOrientArrayType TrigonalLowOps::determineRodriguesVector(uint64_t seed, int cho
   step[0] = TrigDim1StepValue;
   step[1] = TrigDim2StepValue;
   step[2] = TrigDim3StepValue;
-  phi[0] = static_cast<float>(choose % 72);
-  phi[1] = static_cast<float>((choose / 72) % 72);
-  phi[2] = static_cast<float>(choose / (72 * 72));
+  phi[0] = static_cast<int32_t>(choose % 72);
+  phi[1] = static_cast<int32_t>((choose / 72) % 72);
+  phi[2] = static_cast<int32_t>(choose / (72 * 72));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
