+++ Source/OrientationLib/SpaceGroupOps/TrigonalOps.cpp
@@ -347,7 +347,7 @@ FOrientArrayType TrigonalOps::determineEulerAngles(uint64_t seed, int choose)
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TrigDim1InitValue;
@@ -356,9 +356,9 @@ FOrientArrayType TrigonalOps::determineEulerAngles(uint64_t seed, int choose)
   step[0] = TrigDim1StepValue;
   step[1] = TrigDim2StepValue;
   step[2] = TrigDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
 
@@ -398,7 +398,7 @@ FOrientArrayType TrigonalOps::determineRodriguesVector(uint64_t seed, int choose
 {
   float init[3];
   float step[3];
-  float phi[3];
+  int32_t phi[3];
   float h1, h2, h3;
 
   init[0] = TrigDim1InitValue;
@@ -407,9 +407,9 @@ FOrientArrayType TrigonalOps::determineRodriguesVector(uint64_t seed, int choose
   step[0] = TrigDim1StepValue;
   step[1] = TrigDim2StepValue;
   step[2] = TrigDim3StepValue;
-  phi[0] = static_cast<float>(choose % 36);
-  phi[1] = static_cast<float>((choose / 36) % 36);
-  phi[2] = static_cast<float>(choose / (36 * 36));
+  phi[0] = static_cast<int32_t>(choose % 36);
+  phi[1] = static_cast<int32_t>((choose / 36) % 36);
+  phi[2] = static_cast<int32_t>(choose / (36 * 36));
 
   _calcDetermineHomochoricValues(seed, init, step, phi, choose, h1, h2, h3);
   FOrientArrayType ho(h1, h2, h3);
